+++ src/log.ts
@@ -3,7 +3,7 @@ import * as vscode from 'vscode'
 export type LogLevel = 'off' | 'info' | 'debug';
 
 // ---------- Logger ---------- //
-export class Log {
+export class Logger {
   private static ch = vscode.window.createOutputChannel('Busted Test Explorer')
 
   private static level (): LogLevel {
@@ -20,10 +20,16 @@ export class Log {
   }
 
   static debug (msg: string) {
+    this.developerLog(msg)
     if (this.level() === 'debug') { this.ch.appendLine(`[DEBUG ${new Date().toISOString()}] ${msg}`) }
   }
 
   static error (msg: string) {
+    this.developerLog(msg)
     this.ch.appendLine(`[ERROR ${new Date().toISOString()}] ${msg}`)
+  }
+
+  static developerLog (msg: string, ...optionalParams: any[]) {
+    console.log(msg, optionalParams)
   }
 }